+++ payouts/payouts.go
@@ -17,11 +17,11 @@ import (
 
 // FinancialPayout is a schema definition.
 type FinancialPayout struct {
-	Amount   *float64 `json:"amount,omitempty"`
+	Amount   *float32 `json:"amount,omitempty"`
 	Currency *string  `json:"currency,omitempty"`
 	// Format: date
 	Date            *datetime.Date         `json:"date,omitempty"`
-	Fee             *float64               `json:"fee,omitempty"`
+	Fee             *float32               `json:"fee,omitempty"`
 	ID              *int                   `json:"id,omitempty"`
 	Reference       *string                `json:"reference,omitempty"`
 	Status          *FinancialPayoutStatus `json:"status,omitempty"`
